Update README.md
Sitting at home, coming up with projects for students to work on over the summer and one student would like to learn how to create web-pages, so ... this is born.  No idea what I will do with it, but perhaps there will be interesting things that show up on it.